--- src/main/java/domain/lotto/number/LottoNumberMaker.java
@@ -1,9 +1,0 @@
-package domain.lotto.number;
-
-import java.util.List;
-
-public class LottoNumberMaker {
-    public List<Integer> makeNumbers(final NumberGeneratable numberGeneratable) {
-        return numberGeneratable.generate();
-    }
-}
