--- src/main/java/domain/lotto/number/NumberGeneratable.java
@@ -1,8 +1,0 @@
-package domain.lotto.number;
-
-import java.util.List;
-
-@FunctionalInterface
-public interface NumberGeneratable {
-    List<Integer> generate();
-}
