+++ src/main/java/domain/lotto/ticket/LottoTickets.java
@@ -1,7 +1,5 @@
 package domain.lotto.ticket;
 
-import domain.lotto.number.NumberGeneratable;
-
 import java.util.List;
 
 public class LottoTickets {
